requirements.txt
djangorestframework==3.12.2
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
drfdocs==0.0.11
Django==3.2
swagger-spec-validator==2.7.3
python-decouple
pymysql==0.10.0
mysqlclient==1.4.6
django-filter==2.4.0
django-cors-headers==3.5.0
coreapi==2.3.3
coreschema==0.0.4